+++ src/main/java/main/model/service/PaymentGraphicService.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.UUID;
 
 public interface PaymentGraphicService {
-    void setPaymentGraphic(UUID creditOfferId, float interestRate, double loanSum, int loanTerm);
+    void setPaymentGraphic(UUID creditOfferId, double interestRate, double loanSum, int loanTerm);
 
     List<PaymentGraphic> getAllGraphicsForCreditOffer(UUID creditOfferId);
